Feature: IngamePlayerInfo

+++ Assets/Scripts/Mine/Mine.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class Mine : MonoBehaviour
 {
-    public float damage = 30f;
+    public int damage = 30;
     public float pushForce = 10f;
     public GameObject explosionEffect;
     private bool isTriggered = false;
